+++ ext/truth/java/androidx/test/ext/truth/os/PersistableBundleSubject.java
@@ -21,6 +21,8 @@ import androidx.annotation.RequiresApi;
 import com.google.common.truth.FailureMetadata;
 import com.google.common.truth.Subject;
 import com.google.common.truth.Truth;
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Subject for making assertions about {@link PersistableBundle}s.
@@ -32,6 +34,43 @@ public final class PersistableBundleSubject extends BaseBundleSubject {
 
   public static PersistableBundleSubject assertThat(PersistableBundle persistableBundle) {
     return Truth.assertAbout(persistableBundles()).that(persistableBundle);
+  }
+
+  public final void isEqualTo(PersistableBundle other) {
+    // If either are null, just fall back to an equals() comparison.
+    if (actual == null || other == null) {
+      super.isEqualTo(other);
+      return;
+    }
+
+    // PersistableBundle doesn't implement equals() so we convert it to a Map so we can use the
+    // existing Truth implementation.
+    Truth.assertThat(toMap(actual)).isEqualTo(toMap(other));
+  }
+
+  public final void isNotEqualTo(PersistableBundle other) {
+    // If either are null, just fall back to an equals() comparison.
+    if (actual == null || other == null) {
+      super.isNotEqualTo(other);
+      return;
+    }
+
+    // PersistableBundle doesn't implement equals() so we convert it to a Map so we can use the
+    // existing Truth implementation.
+    Truth.assertThat(toMap(actual)).isNotEqualTo(toMap(other));
+  }
+
+  /** Converts the {@link PersistableBundle} to a {@link Map} for comparison. */
+  private final Map<String, Object> toMap(PersistableBundle bundle) {
+    Map<String, Object> map = new HashMap<>();
+    for (String key : bundle.keySet()) {
+      Object value = bundle.get(key);
+      if (value instanceof PersistableBundle) {
+        value = toMap((PersistableBundle) value);
+      }
+      map.put(key, value);
+    }
+    return map;
   }
 
   public static Subject.Factory<PersistableBundleSubject, PersistableBundle> persistableBundles() {